Interpolate population data faster
Makes use of pandas primitives, avoiding python functions and loops.

Also kinda changes the format of the data. Each state is a column now,
so each column represents a time series. The time range is expanded
to the min and max of all the data, so there are plently of chunks of
time where there are no reports. Probably you want to shrink the range.

Will, you'll probably need to modify the later code if you want to use this,
I started trying but it was annoying cuz I didn't understand how the
plotting work. You do though so should be faster for you.