Abstract out the signature of tweak_style.py
The script had been adjusting the style of the rates.html doc only. Now
it's set up to recieve any html doc generated by the jupyter notebook
and format it according to a few rules. Additionally, I pulled out the
font-family as its own variable so that I can adjust it for the legend
and the slider output all at once.